Adds tests for ignoring invalid ARN
